Made composer package out of saplo php library
Added namespacing and correct psr4 autloading
